Move RulesVerifier to the RulesEngine namespace
Also moves RuleStatus used by the Verifier to the RuleEngine namespace.
